refactor: modularize SoftDevice
